requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
faiss-cpu==1.9.0
openai==1.55.3
pypdf==5.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.40.2
streamlit-extras==0.5.0
streamlit-vertical-slider==2.5.5
tiktoken==0.8.0
Flask==3.1.0
Werkzeug==3.1.3